backend/requirements.txt
Flask
tensorflow
numpy
transformers
torch
gunicorn
Flask-CORS
opencv-python
tf-keras
requests
beautifulsoup4
requests-html
lxml
discord.py
psutil
onnx
onnxruntime
torch
pandas
scikit-learn
Flask-SQLAlchemy
python-nmap
scapy
getmac
prettytable
Flask-Migrate
PyNaCl
pyppeteer==1.0.0
nest_asyncio
gpt4all
lxml_html_clean
